AltitudeHeightConverter: use float for floats
 * do not try to store float values in an int

+++ include/fimex/coordSys/verticalTransform/AltitudeHeightConverter.h
@@ -22,7 +22,7 @@ private:
     VerticalConverterPtr altitudeOrHeight_;
     std::string topography_;
     std::string topographyUnit_;
-    int topographyFactor_;
+    float topographyFactor_;
 };
 
 } // namespace MetNoFimex
